% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_exif.R
\name{apply_gps_to_images}
\alias{apply_gps_to_images}
\title{Apply GPS to Images}
\usage{
apply_gps_to_images(csv_path, photo_dir)
}
\arguments{
\item{csv_path}{Path to the CSV manifest created by write_gps_manifest().}

\item{photo_dir}{Path to the directory containing images.}
}
\value{
NULL. Prints output to console.
}
\description{
Wrapper for the ExifTool system command.
Requires 'exiftool' to be in the system PATH.
}
