% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_data.R
\name{get_photo_timestamps}
\alias{get_photo_timestamps}
\title{Get Photo Timestamps}
\usage{
get_photo_timestamps(folder_path, tz_offset = "Pacific/Honolulu")
}
\arguments{
\item{folder_path}{String. Path to the folder containing images.}

\item{tz_offset}{String. Timezone of the camera clock (e.g., "Pacific/Honolulu").}
}
\value{
A tibble with columns: SourceFile, local_time, timestamp_utc.
}
\description{
Scans a directory for image files and extracts their "DateTimeOriginal" tags.
}
