% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_locations.R
\name{match_semantic_places}
\alias{match_semantic_places}
\title{Match Semantic Places (Plan B)}
\usage{
match_semantic_places(photo_df, semantic_df)
}
\arguments{
\item{photo_df}{Dataframe of photos with timestamp_utc.}

\item{semantic_df}{Dataframe of visits with start_utc and end_utc.}
}
\value{
A tibble with lat_sem and lon_sem.
}
\description{
Matches photos to a "Place Visit" based on time overlap.
}
