% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timelinesr-package.R
\docType{package}
\name{timelinesr-package}
\alias{timelinesr}
\alias{timelinesr-package}
\title{timelinesr: Adds Locations from an Android Timeline to Images}
\description{
Use timeline location data captured with a cellphone. Both detailed (last 30 day) and summarized values (beyond 30 day) are used to approximate the photo locations done with time comparisons. Locations are stored in the photo EXIF data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kimbridges/timelinesr}
  \item Report bugs at \url{https://github.com/kimbridges/timelinesr/issues}
}

}
\author{
\strong{Maintainer}: Kim Bridges \email{kim.bridges@gmail.com}

}
\keyword{internal}
