% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{write_geotags}
\alias{write_geotags}
\title{Write Geotags to Image Files}
\usage{
write_geotags(tagged_df)
}
\arguments{
\item{tagged_df}{The dataframe returned by geotag_photos().}
}
\description{
Uses ExifTool to write the calculated latitude and longitude back into the
original image files.
}
