% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_geotags}
\alias{visualize_geotags}
\title{Visualize Geotagged Photos}
\usage{
visualize_geotags(tagged_df)
}
\arguments{
\item{tagged_df}{The dataframe returned by geotag_photos().}
}
\description{
Creates a simple map of where photos were taken.
}
