% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{write_geotags}
\alias{write_geotags}
\title{Write Geotags to Files}
\usage{
write_geotags(tagged_df, photo_folder)
}
\arguments{
\item{tagged_df}{The dataframe returned by geotag_photos().}

\item{photo_folder}{String. The folder where images are located.}
}
\description{
A wrapper that generates the manifest and immediately runs ExifTool.
}
